+++ app/src/main/java/com/wardrob/wardrob/weather/model/Wind.java
@@ -1,8 +1,6 @@
 package com.wardrob.wardrob.weather.model;
 
-/**
- * Created by stas.savinov on 16.02.2018.
- */
+
 public class Wind
 {
     private double speed;
@@ -11,9 +9,9 @@ public class Wind
 
     public void setSpeed(double speed) { this.speed = speed; }
 
-    private int deg;
+    private double deg;
 
-    public int getDeg() { return this.deg; }
+    public double getDeg() { return this.deg; }
 
-    public void setDeg(int deg) { this.deg = deg; }
+    public void setDeg(double deg) { this.deg = deg; }
 }
